Fixed am inlining bug causing an error for procedures that contains assignments where the local variable (which is renamned in the inlining) appeared both at the LHS and RHS (e.g. x := x + 1)

+++ trunk/core/source/java/net/sf/opendf/cli/SSAGenerator.java
@@ -269,7 +269,7 @@ public class SSAGenerator extends XSLTTransformRunner
             try
             {
                 final Node ecalmlNode = Elaborator.elaborateModel(this.configs, null, SSAGenerator.class.getClassLoader());
-                writeFile(new File(this.getRunDir(), prefix+".ecalml"), Util.createXML(ecalmlNode));
+                if (this.isPreserveFiles()) writeFile(new File(this.getRunDir(), prefix+".ecalml"), Util.createXML(ecalmlNode));
                 calmlNode = Elaborator.elabPostProcess(ecalmlNode, this.configs, null, SSAGenerator.class.getClassLoader());
             }catch (Exception e ){
                 throw new SubProcessException("Could not elaborate top model",e);
